Last cleaning for comments and touch split
