+++ nanoFramework.Graphics/nanoFramework.Graphics.nfproj
@@ -237,13 +237,20 @@
     <Compile Include="Primitive\SpiConfiguration.cs" />
     <Compile Include="Primitive\Touch\CollectionMethod.cs" />
     <Compile Include="Primitive\Touch\CollectionMode.cs" />
+    <Compile Include="Primitive\Touch\Touch.cs" />
     <Compile Include="Primitive\Touch\TouchCollector.cs" />
     <Compile Include="Primitive\Touch\TouchCollectorConfiguration.cs" />
+    <Compile Include="Primitive\Touch\TouchEvent.cs" />
+    <Compile Include="Primitive\Touch\TouchEventProcessor.cs" />
     <Compile Include="Primitive\Touch\TouchEvents.cs" />
+    <Compile Include="Primitive\Touch\TouchGesture.cs" />
+    <Compile Include="Primitive\Touch\TouchGestureEventArgs.cs" />
+    <Compile Include="Primitive\Touch\TouchInput.cs" />
     <Compile Include="Primitive\Touch\TouchInputFlags.cs" />
-    <Compile Include="Primitive\Touch\TouchInterface.cs" />
+    <Compile Include="Primitive\Touch\TouchMessages.cs" />
     <Compile Include="Primitive\Touch\TouchPanel.cs" />
     <Compile Include="Primitive\Touch\TouchScreen.cs" />
+    <Compile Include="Primitive\Touch\TouchScreenEventArgs.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="System\Application.cs" />
     <Compile Include="System\EventHandler.cs" />
